feat: allow non-base64 encoded payloads
If the sender sends application/json, then the Lambda system does _not_ base64
encode the body. So to get the original bytes, we need to encode the payload.

(Technically if the payload was encoded using something other than utf-8, this
won't be the original bytes, but for our purposes we can live with that)